% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRISMA_flow.R
\name{insertJS}
\alias{insertJS}
\title{Insert rotated text to side panel nodes}
\usage{
insertJS(plot)
}
\arguments{
\item{plot}{A PRISMA_flowchart plot.}
}
\value{
A plot object containing javascript code.
}
\description{
Text cannot be rotated within DiagrammeR, so this function appends
a Javascript function to add in rotated text after the id="node" 'g' tag that
corresponds to the blue nodes at the side of the plot.
}
\examples{
\dontrun{
plot <- insertJS(plot);
plot;
}
}
