% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PRISMA_add_hyperlink_}
\alias{PRISMA_add_hyperlink_}
\title{Add the hyperlink to the given node}
\usage{
PRISMA_add_hyperlink_(node, url)
}
\arguments{
\item{node}{the relevent node}

\item{url}{the URL the node should link to}
}
\value{
An interactive flow diagram plot.
}
\description{
Generate the javascript method to insert the side labels
}
\keyword{internal}
