% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PRISMA_calc_filetype_}
\alias{PRISMA_calc_filetype_}
\title{Calculate the correct filetime}
\usage{
PRISMA_calc_filetype_(fn, ft)
}
\arguments{
\item{fn}{The filename (including extension)}

\item{ft}{The filetype (which can be NA or NULL)}
}
\value{
the filetype taken from the filename, or overriden by the ft param
}
\description{
Work out the correct filetype to save the file as
}
\keyword{internal}
