% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRISMA_checklist.R
\name{PRISMA_checklistdata}
\alias{PRISMA_checklistdata}
\title{Save PRISMA2020 flow diagram}
\usage{
PRISMA_checklistdata(filename, filetype, h = 8, w = 11, fontsize = 6)
}
\arguments{
\item{filename}{The filename of the checklist (including extension)}

\item{h}{Height of the page (default to just below A4 height)}

\item{w}{Width of the page (default to just below A4 height)}

\item{fontsize}{Fontsize (default to 6)}

\item{plotobj}{A plot produced using \code{\link[=PRISMA_flowdiagram]{PRISMA_flowdiagram()}}.}

\item{overwrite}{if TRUE, will overwrite an existing file}

\item{width}{passed as the width argument to
\code{\link[rsvg:rsvg]{rsvg::rsvg()}} and similar functions}

\item{height}{passed as the height argument to
\code{\link[rsvg:rsvg]{rsvg::rsvg()}} and similar functions}

\item{css}{passed as the css argument to
\code{\link[rsvg:rsvg]{rsvg::rsvg()}} and similar functions}
}
\value{
the absolute filename of the saved diagram plot.
}
\description{
Save the output from \code{\link[=PRISMA_flowdiagram]{PRISMA_flowdiagram()}} to the
working directory.
}
