% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PRISMA_default_or_csv_}
\alias{PRISMA_default_or_csv_}
\title{Pull from CSV, or use a default value if not in the CSV}
\usage{
PRISMA_default_or_csv_(expr, default, var_name = "not provided")
}
\arguments{
\item{expr}{the expression to get the variable from the CSV}

\item{default}{the default token}

\item{var_name}{the variable name}
}
\value{
the result of 'expr', or 'default'
}
\description{
set the variable from the CSV, or a default value
}
\keyword{internal}
