% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRISMA_flowdiagram.R
\name{PRISMA_export_csv}
\alias{PRISMA_export_csv}
\title{Export the input data object as a CSV}
\usage{
PRISMA_export_csv(
  data,
  filename = "PRISMA2020_flowdiagram.csv",
  overwrite = FALSE
)
}
\arguments{
\item{data}{A data object produced by \code{\link[=PRISMA_flowdiagram]{PRISMA_flowdiagram()}}.}

\item{filename}{The filename to save (including extension)}

\item{overwrite}{if TRUE, will overwrite an existing file}
}
\value{
the absolute filename of the saved CSV.
}
\description{
Save the output from \code{\link[=PRISMA_data]{PRISMA_data()}} to a CSV.
}
\examples{
csvFile <- system.file("extdata", "PRISMA.csv", package = "PRISMA2020")
data <- read.csv(csvFile);
data <- PRISMA_data(data);
PRISMA_export_csv(data, filename = tempfile());
}
