% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRISMA_flowchart.R
\name{PRISMA_flowchart}
\alias{PRISMA_flowchart}
\title{Plot interactive flow charts for systematic reviews}
\usage{
PRISMA_flowchart(
  previous_studies,
  previous_reports,
  register_results,
  database_results,
  website_results,
  organisation_results,
  citations_results,
  duplicates,
  excluded_automatic,
  excluded_other,
  records_screened,
  records_excluded,
  dbr_sought_reports,
  dbr_notretrieved_reports,
  other_sought_reports,
  other_notretrieved_reports,
  dbr_assessed,
  dbr_excluded,
  other_assessed,
  other_excluded,
  new_studies,
  new_reports,
  total_studies,
  total_reports,
  interactive = FALSE,
  tooltips = "",
  urls = "",
  font = "Helvetica",
  title_colour = "Goldenrod1",
  greybox_colour = "Gainsboro",
  main_colour = "Black",
  arrow_colour = "Black",
  arrow_head = "normal",
  arrow_tail = "none"
)
}
\arguments{
\item{previous_studies}{The number of studies included in previous version of
review.}

\item{previous_reports}{The number of reports of studies included in previous
version of review.}

\item{register_results}{The number of records identified from registers.}

\item{database_results}{The number of records identified from databases.}

\item{website_results}{The number of records identified from websites.}

\item{organisation_results}{The number of records identified from
organisations.}

\item{citations_results}{The number of records identified from citation
searching.}

\item{duplicates}{The number of duplicate records.}

\item{excluded_automatic}{The number of records marked as ineligible by
automation tools.}

\item{excluded_other}{The number of records removed for other reasons.}

\item{records_screened}{The number of records screened for databases and
registers.}

\item{records_excluded}{The number of records excluded for databases and
registers.}

\item{dbr_sought_reports}{The number of reports sought for retrieval for
databases and registers.}

\item{dbr_notretrieved_reports}{The number of reports not retrieved for
databases and registers.}

\item{other_sought_reports}{The number of reports sought for retrieval
from other sources.}

\item{other_notretrieved_reports}{The number of reports not retrieved for
retrievalfrom other sources.}

\item{dbr_assessed}{The number of reports assessed for eligibility for
databases and registers.}

\item{dbr_excluded}{The number of reports excluded for databases and
registers: (separate reasons and numbers using ; e.g. Reason1, xxx;
Reason2, xxx; Reason3, xxx).}

\item{other_assessed}{The number of reports assessed for eligibility for
other sources.}

\item{other_excluded}{The number of reports excluded for other sources:
(separate reasons and numbers using ; e.g. Reason1, xxx; Reason2, xxx;
Reason3, xxx).}

\item{new_studies}{The number of new studies included in review.}

\item{new_reports}{The number of reports of new included studies.}

\item{total_studies}{The number of total studies included in review.}

\item{total_reports}{The number of total reports included studies.}

\item{interactive}{Logical argument TRUE or FALSE whether to plot interactivity
(tooltips and hyperlinked boxes).}

\item{tooltips}{Mouseover popups for each box containing explanatory text.
Should be provided as a vector.}

\item{urls}{A dataframe of urls to act as hyperlinks for each box, with one column
(named 'box') corresponding to the boxname ('box1', etc.) and one column (named
'url') containing the urls. A total of 16 URLs should be provided, one for each box.
See the 'PRISMA_flow_schema.png' for details of box numbers, sequentially from
top left to bottom across columns, from left to right. Additional URLs can be given
for the side and top rounded panel boxes if desired.}

\item{font}{The font for text in each box. The default is 'Helvetica'.}

\item{title_colour}{The colour for the upper middle title box (new studies).
The default is 'Goldenrod1'. See DiagrammeR colour scheme
\url{http://rich-iannone.github.io/DiagrammeR/graphviz_and_mermaid.html#colors}.}

\item{greybox_colour}{The colour for the left and right column boxes. The
default is 'Gainsboro'. See DiagrammeR colour scheme
\url{http://rich-iannone.github.io/DiagrammeR/graphviz_and_mermaid.html#colors}.}

\item{main_colour}{The colour for the main box borders. The default is
'Black'. See DiagrammeR colour scheme
\url{http://rich-iannone.github.io/DiagrammeR/graphviz_and_mermaid.html#colors}.}

\item{arrow_colour}{The colour for the connecting lines. The default
is 'Black'. See DiagrammeR colour scheme
\url{http://rich-iannone.github.io/DiagrammeR/graphviz_and_mermaid.html#colors}.}

\item{arrow_head}{The head shape for the line connectors. The default is
'normal'. See DiagrammeR arrow shape specification
\url{http://rich-iannone.github.io/DiagrammeR/graphviz_and_mermaid.html#arrow-shapes}.}

\item{arrow_tail}{The tail shape for the line connectors. The default is
'none'. See DiagrammeR arrow shape specification
\url{http://rich-iannone.github.io/DiagrammeR/graphviz_and_mermaid.html#arrow-shapes}.}
}
\value{
A flow chart plot.
}
\description{
Produces a PRISMA2020 style flow chart for systematic reviews,
with the option to add interactivity through tooltips (mouseover popups) and
hyperlink URLs to each box. Data can be imported from the standard CSV template
provided.
}
\examples{
\dontrun{
data <- read.csv(file.choose());
data <- read_PRISMAdata(data);
attach(data);
plot <- PRISMA_flowchart(previous_studies = previous_studies,
                previous_reports = previous_reports,
                register_results = register_results,
                database_results = database_results,
                website_results = website_results,
                organisation_results = organisation_results,
                citations_results = citations_results,
                duplicates = duplicates,
                excluded_automatic = excluded_automatic,
                excluded_other = excluded_other,
                records_screened = records_screened,
                records_excluded = records_excluded,
                dbr_sought_reports = dbr_sought_reports,
                dbr_notretrieved_reports = dbr_notretrieved_reports,
                other_sought_reports = other_sought_reports,
                other_notretrieved_reports = other_notretrieved_reports,
                dbr_assessed = dbr_assessed,
                dbr_excluded = dbr_excluded,
                other_assessed = other_assessed,
                other_excluded = other_excluded,
                new_studies = new_studies,
                new_reports = new_reports,
                total_studies = total_studies,
                total_reports = total_reports,
                interactive = TRUE,
                tooltips = tooltips,
                urls = urls);
plot
}
}
