% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PRISMA_format_number_}
\alias{PRISMA_format_number_}
\title{Format numbers with commas into numbers}
\usage{
PRISMA_format_number_(x)
}
\arguments{
\item{x}{the number to format}
}
\value{
the number with commas removed
}
\description{
Turn strings containing numbers +/- commas into numbers
}
\keyword{internal}
