% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PRISMA_format_reasons_}
\alias{PRISMA_format_reasons_}
\title{Formats multiple exclusion reasons properly for printing}
\usage{
PRISMA_format_reasons_(df)
}
\arguments{
\item{df}{the dataframe to parse}
}
\value{
a string ready for printing
}
\description{
Parse an exclusion reason dataframe from
\code{\link[=PRISMA_parse_reasons_]{PRISMA_parse_reasons_()}} and returns a properly formatted string
}
\keyword{internal}
