% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PRISMA_gen_tmp_svg_}
\alias{PRISMA_gen_tmp_svg_}
\title{Generate a temporary SVG from a plot object}
\usage{
PRISMA_gen_tmp_svg_(obj)
}
\arguments{
\item{obj}{the plot object}
}
\value{
the full path to the saved SVG
}
\description{
Generate and save a temporary SVG from a plot object
}
\keyword{internal}
