% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PRISMA_get_pos_}
\alias{PRISMA_get_pos_}
\title{Calculate the correct position of a node}
\usage{
PRISMA_get_pos_(first_box_location, offset, length_orig, length_new)
}
\arguments{
\item{first_box_location}{the location of the first node}

\item{offset}{the offset from the first node}

\item{length_orig}{the width/height of the original node}

\item{length_new}{the width/height of the new node}
}
\value{
the position of the node
}
\description{
Get the correct position for a node
}
\keyword{internal}
