% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PRISMA_parse_reasons_}
\alias{PRISMA_parse_reasons_}
\title{Parse an exclusion reason into a data frame}
\usage{
PRISMA_parse_reasons_(reasons)
}
\arguments{
\item{reasons}{the string to parse}
}
\value{
a dataframe containing reasons and number applicable
}
\description{
Parse an exclusion reason string and returns a dataframe
containing reasons and number
}
\keyword{internal}
