% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRISMA_flowdiagram.R
\name{read_PRISMAdata}
\alias{read_PRISMAdata}
\title{Read in PRISMA flow diagram data}
\usage{
read_PRISMAdata(data)
}
\arguments{
\item{data}{File to read in.}
}
\value{
A list of objects needed to plot the flow diagram
}
\description{
Read in a template CSV containing data for the flow diagram
}
\examples{
\dontrun{
data <- read.csv(file.choose(), stringsAsFactors=FALSE);
data <- read_PRISMAdata(data);
attach(data);
}
}
