% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PRISMA_get_height_}
\alias{PRISMA_get_height_}
\title{Calculate the correct height of a box from a list (e.g. of exclusion reasons)}
\usage{
PRISMA_get_height_(list)
}
\arguments{
\item{list}{a correctly wrapped (using stringr) list}
}
\value{
the height of the box
}
\description{
Get the correct height for a box
}
\keyword{internal}
