% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PRISMA_interactive_}
\alias{PRISMA_interactive_}
\title{Plot interactive flow diagram for systematic reviews}
\usage{
PRISMA_interactive_(plot, urls, previous, other)
}
\arguments{
\item{plot}{A plot object from \code{\link[=PRISMA_flowdiagram]{PRISMA_flowdiagram()}}.}

\item{urls}{A dataframe consisting of two columns: nodes and urls. The first
column should contain 19 rows for the nodes from node1 to node19. The second
column should contain a corresponding URL for each node.}

\item{previous}{Logical argument (TRUE or FALSE) (supplied through
\code{\link[=PRISMA_flowdiagram]{PRISMA_flowdiagram()}}) specifying whether previous studies were sought.}

\item{other}{Logical argument (TRUE or FALSE) (supplied through
\code{\link[=PRISMA_flowdiagram]{PRISMA_flowdiagram()}}) specifying whether other studies were sought.}
}
\value{
An interactive flow diagram plot.
}
\description{
Converts a PRISMA systematic review flow diagram into an
interactive HTML plot, for embedding links from each box.
}
\seealso{
\code{\link[=PRISMA_interactive_]{PRISMA_interactive_()}}
}
\keyword{internal}
