% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRISMA_flowdiagram.R
\name{PRISMA_save}
\alias{PRISMA_save}
\title{Save PRISMA2020 flow diagram}
\usage{
PRISMA_save(plotobj, format = "HTML", filename = "PRISMA2020_flowdiagram")
}
\arguments{
\item{plotobj}{A plot produced using PRISMA_flowdiagram().}

\item{format}{The format to save the plot in, supports: HTML, PDF, PNG, SVG, PS and WEBP}

\item{filename}{The filename to save (without extension)}
}
\value{
A flow diagram plot as an html file, with embedded links and
tooltips if interactive=TRUE in PRISMA_flowdiagram() and if tooltips
are provided in the data upload, respectively.
}
\description{
Save the output from PRISMA_flowdiagram() to the
working directory.
}
\examples{
\dontrun{
data <- read.csv(file.choose());
data <- read_PRISMAdata(data);
attach(data); 
plot <- PRISMA_flowdiagram(data,
                fontsize = 12,
                interactive = TRUE,
                previous = TRUE,
                other = TRUE)
PRISMA_save(plot, format = 'pdf')
}
}
