% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRISMA_flowdiagram.R
\name{PRISMA_save}
\alias{PRISMA_save}
\title{Save PRISMA2020 flow diagram}
\usage{
PRISMA_save(plotobj)
}
\arguments{
\item{plotobj}{A plot produced using PRISMA_flowdiagram().}
}
\value{
A flow diagram plot as an html file, with embedded links and
tooltips if interactive=TRUE in PRISMA_flowdiagram() and if tooltips
are provided in the data upload, respectively.
}
\description{
Save the html output from PRISMA_flowdiagram() to the
working directory.
}
