% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRISMA_flowdiagram.R
\name{read_PRISMAdata}
\alias{read_PRISMAdata}
\title{Defunct function - replaced by "PRISMA_data"}
\usage{
read_PRISMAdata(data)
}
\arguments{
\item{data}{File to read in.}
}
\description{
Defunct function - replaced by "PRISMA_data"
}
\seealso{
\code{\link[=PRISMA_data]{PRISMA_data()}}
}
